модулируем код ч 1
- разнесены по разным файлам основные константы и функции
- сборка протестирована - работает, ошибок нет, функциональность созранена
